Add a pg_anonymize.check_labels option.
This allows to perform sanity checks when declaring a SECURITY LABEL:

- check that the given expression is valid when used on the target table
- check that the given expression returns the same data type as the target
  column

If the target column type is text, we allow "unknown" datatype as this is
usually ok, but we still notify the user and hint to explicitly cast the
expression to text.

Also do a mandatory check that the given expression doesn't contain any SQL
injection attempt.