Improve performance of relation anonymization.
Previous code was calling GetSecurityLabel, which itself does an index scan on
pg_seclabel, for each attribute of the relation.  This is quite inefficient as
you have to do O(n) index scan, even if the relation isn't anonymized at all.

Instead, simply retrieve all security labels for the given relation in a single
index scan and store those in an array indexed by the attribute's AttrNumber if
any security label is found, or simply notify the caller that the relation
isn't anonymized.