Fix COPY TO behavior with anonymized partitioned / inheritance tables.
COPY TO isn't allowed for partitioned table, so we can just ignore those when
asked to generate a query for a given relation.

For inheritance tree, COPY TO behavior is to only handle the current relation
and not any inherited table (so the same behavior as FROM ONLY).  Fix the
behavior by using a FROM ONLY in the generated query in that case.

Regression tests updated.  While at it consitently use uppercase STDOUT.