Try to limit side effects with other plugins.
This extension can modify the Query in post_parse_analyze_hook, but
doesn't have a way to adapt the raw query string accordingly.  This can
cause problem with some extensions like pg_stat_statements that rely on
both referring to the same thing.  To make sure that we don't cause
interference, we process other post_parse_analyze_hook first before our
own processing, but we have to make sure that we're the last module
loaded.  Unfortunately we can only check that when our code is loaded,
so we can only hope that no incompatible extension will be loaded
afterwards.